rename llama-stack e2e and prow artifacts to ogx
