+++ tests/e2e-prow/rhoai/manifests/lightspeed/ogx-prow.yaml
