+++ tests/e2e/utils/README.md
@@ -1,14 +1,14 @@
 # List of source files stored in `tests/e2e/utils` directory
 
-## [llama_config_utils.py](llama_config_utils.py)
+## [ogx_config_utils.py](ogx_config_utils.py)
 
 Helpers for reading and updating OGX run.yaml across environments.
 
-## [llama_prow_utils.py](llama_prow_utils.py)
+## [ogx_prow_utils.py](ogx_prow_utils.py)
 
 Thin Prow/OpenShift wrappers for OGX run.yaml ConfigMap operations.
 
-## [llama_stack_utils.py](llama_stack_utils.py)
+## [ogx_utils.py](ogx_utils.py)
 
 E2E test utilities for OGX shields.
